separate out DummyFs class from dummy_file_oo.php
